requirements.txt
Django
mysqlclient
djangorestframework
django-filter